task 2/7: Extract Merge, UnMerge, Switch, and TransferIbc builders into per-option DepositSubmitStrategy files (#4862)
